Implement intake and extraction services with validation and parsing stubs
