+++ packages/api/src/services/extraction-service.ts
